refactor: extract hover card state logic into hook
