+++ packages/web/src/modules/glossary-item-info/component.tsx
@@ -1,7 +1,7 @@
 import type { BlocksContent } from "@strapi/blocks-react-renderer";
 import BlocksRenderer from "@/modules/strapi/blocks-renderer";
 import { HoverCard, HoverCardContent, HoverCardTrigger } from "@/components/ui/hover-card";
-import { useState } from "react";
+import { useHoverCardState } from "@/modules/common/use-hover-card-state";
 
 type GlossaryItemInfoComponentProps = {
   name: string;
@@ -14,18 +14,11 @@ export function GlossaryItemInfoComponent({
   description,
   onOpenChange,
 }: GlossaryItemInfoComponentProps) {
-  const [isCardOpen, setIsCardOpen] = useState(false);
-
-  const handleOpenChange = (open: boolean) => {
-    setIsCardOpen(open);
-    if (onOpenChange) {
-      onOpenChange(open);
-    }
-  };
+  const { isCardOpen, handleOpenChange, toggleOpen } = useHoverCardState(onOpenChange);
 
   return (
     <HoverCard openDelay={150} open={isCardOpen} onOpenChange={handleOpenChange}>
-      <HoverCardTrigger asChild onClick={() => handleOpenChange(!isCardOpen)}>
+      <HoverCardTrigger asChild onClick={toggleOpen}>
         <span className="underline cursor-help">{name}</span>
       </HoverCardTrigger>
       <HoverCardContent
